+++ test/functional/rpc_deprecated.py
@@ -9,7 +9,7 @@ class DeprecatedRpcTest(BitcoinTestFramework):
     def set_test_params(self):
         self.num_nodes = 2
         self.setup_clean_chain = True
-        self.extra_args = [[], []]
+        self.extra_args = [[], ["-deprecatedrpc=validateaddress"]]
 
     def run_test(self):
         # This test should be used to verify correct behaviour of deprecated
@@ -18,10 +18,13 @@ class DeprecatedRpcTest(BitcoinTestFramework):
         # self.log.info("Make sure that -deprecatedrpc=createmultisig allows it to take addresses")
         # assert_raises_rpc_error(-5, "Invalid public key", self.nodes[0].createmultisig, 1, [self.nodes[0].getnewaddress()])
         # self.nodes[1].createmultisig(1, [self.nodes[1].getnewaddress()])
-        #
-        # There are currently no deprecated RPC methods in master, so this
-        # test is currently empty.
-        pass
+
+        self.log.info("Test validateaddress deprecation")
+        SOME_ADDRESS = "mnvGjUy3NMj67yJ6gkK5o9e5RS33Z2Vqcu" # This is just some random address to pass as a parameter to validateaddress
+        dep_validate_address = self.nodes[0].validateaddress(SOME_ADDRESS)
+        assert "ismine" not in dep_validate_address
+        not_dep_val = self.nodes[1].validateaddress(SOME_ADDRESS)
+        assert "ismine" in not_dep_val
 
 if __name__ == '__main__':
     DeprecatedRpcTest().main()
